+++ src/avatar/story.tsx
@@ -1,23 +1,22 @@
 import React from 'react'
 
 import { storiesOf } from '@storybook/react'
-import { withInfo } from '@storybook/addon-info'
 import { withKnobs, boolean, text, select } from '@storybook/addon-knobs'
 
 import Avatar from 'avatar'
 
 const sizes = {
   default: '',
-  isMedium: 'medium',
-  isLarge: 'large',
+  isMedium: 'isMedium',
+  isLarge: 'isLarge',
 }
 
 const stories = storiesOf('Avatar', module)
 stories.addDecorator(withKnobs)
 
 stories.add('Basic', () => {
-  const modifiers = select('Size', sizes, 'default')
-  const propsModifiers = { [modifiers]: !!modifiers }
+  const modifiers = select('Size', sizes, '')
+  const propsModifiers = modifiers ? { [modifiers]: !!modifiers } : {}
   return (
     <Avatar
       image={text('url', '//placehold.it/80x80')}